+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 ### Added
 - Support of `SimplePredicate` and `DatePredicate` with two attributes.
 
+### Changed
+- Now during deleting record in offline mode detail records are deleted too.
+
 ## [2.7.0-beta.0] - 2021-09-20
 ### Breaking changes
 - Deleted `custom-inflector-rules`. Model names for OData created by adding `s`.